requirements.txt
pip
twine
pytest

numpy
pandas
bs4
selenium

python-dotenv
psutil

sqlalchemy
psycopg2-binary
chromedriver-binary==94.0.4606.81-1
erddapy
podaacpy

netcdf4
h5py
xarray
